requirements.txt
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=4.9.3,<5.0.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pytest>=8.3.2
colorama>=0.4.6
